+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/BetterTensorFlow.kt
@@ -12,8 +12,8 @@ public enum class CupLocation {
     UNKNOWN
 }
 
-@Disabled
 @TeleOp(name="BetterTensorFlow", group="Concept")
+@Disabled
 class BetterTensorFlow: DriveMethods(){
     override fun runOpMode() {
         initVision(Variables.VisionProcessors.TFOD)
