+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Teleop/Testing/ClawController.kt
@@ -1,11 +1,13 @@
 
-package org.firstinspires.ftc.teamcode.Teleop
+package org.firstinspires.ftc.teamcode.Teleop.Testing
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import com.qualcomm.robotcore.hardware.Servo
 import org.firstinspires.ftc.teamcode.DriveMethods
 
 @TeleOp(name = "ClawControl", group = "AprilTag")
+@Disabled
 class ClawControl: DriveMethods() {
     var servo1: Servo? = null;
     var servo2: Servo? = null;
